Jest: Enable rendering tests based on react-testing library. Removed enzyme.
