+++ frontend/jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   testURL: 'http://localhost:8000',
   verbose: false,
-  // extraSetupFiles: ['./tests/setupTests.js'],
+  setupFilesAfterEnv: ['./tests/setupTests.jsx'],
   globals: {
     ANT_DESIGN_PRO_ONLY_DO_NOT_USE_IN_YOUR_PRODUCTION: false,
     localStorage: null,
