+++ frontend/package.json
@@ -108,7 +108,6 @@
     "cross-env": "^7.0.0",
     "cross-port-killer": "^1.3.0",
     "detect-installer": "^1.0.0",
-    "enzyme": "^3.11.0",
     "eslint": "^7.32.0",
     "express": "^4.17.0",
     "gh-pages": "^3.2.0",
